created new branch for game and sort files
