+++ src/App.jsx
@@ -3,7 +3,7 @@ import Header from "./components/Header/Header"; // Adjust the path as needed
 import Footer from "./components/Footer"; // Adjust the path as needed
 import Home from "./components/Home";
 import About from "./components/About";
-import Contact from "./components/Contact";
+import Game from "./components/Game/Game";
 import "./App.css";
 
 function App() {
@@ -13,7 +13,7 @@ function App() {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/about" element={<About />} />
-        <Route path="/contact" element={<Contact />} />
+        <Route path="/game" element={<Game />} />
       </Routes>
       <Footer />
     </Router>
