+++ src/components/Game/GameOver.jsx
