+++ src/components/Game/Log.jsx
