--- src/components/winning-combinations.js
@@ -1,42 +1,0 @@
-export const WINNING_COMBINATIONS = [
-    [
-      { row: 0, column: 0 },
-      { row: 0, column: 1 },
-      { row: 0, column: 2 },
-    ],
-    [
-      { row: 1, column: 0 },
-      { row: 1, column: 1 },
-      { row: 1, column: 2 },
-    ],
-    [
-      { row: 2, column: 0 },
-      { row: 2, column: 1 },
-      { row: 2, column: 2 },
-    ],
-    [
-      { row: 0, column: 0 },
-      { row: 1, column: 0 },
-      { row: 2, column: 0 },
-    ],
-    [
-      { row: 0, column: 1 },
-      { row: 1, column: 1 },
-      { row: 2, column: 1 },
-    ],
-    [
-      { row: 0, column: 2 },
-      { row: 1, column: 2 },
-      { row: 2, column: 2 },
-    ],
-    [
-      { row: 0, column: 0 },
-      { row: 1, column: 1 },
-      { row: 2, column: 2 },
-    ],
-    [
-      { row: 0, column: 2 },
-      { row: 1, column: 1 },
-      { row: 2, column: 0 },
-    ],
-  ];
